requirements.txt
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.9
dagshub-annotation-converter==0.1.9
databricks-sdk==0.53.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
locket==1.0.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mlflow==2.22.0
mlflow-skinny==2.22.0
modin==0.32.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.4
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pathvalidate==3.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
polars==1.27.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.4.0
ray==2.44.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruff==0.11.4
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelib==1.7.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
